idpf-linux: refactor xdp desc prep func to pass params struct
With the changes to the Tx path to use a buffer/tag pool instead of a
ring in flow scheduling mode, the XDP tx_buf selection also needs to use
the refillq to determine what buffer index to use when transmitting.
Move the splitq_params struct to the descriptor prep caller and pass it
as an argument. This allows us to determine the compl_tag independent of
the descriptor ring index.

will be passed to the descriptor build func for singleq, but it will be
unused.

Signed-off-by: Joshua Hay <joshua.a.hay@intel.com>
Change-type: Other