idpf-linux: refactor exception path handling for buffer pool
Refactor the expection path (double completion) completion type handling
to work with the buffer pool instead of the old stashing mechanisms.

Use an xarray per queue to effectively "stash" packets that hit the
exception path in the miss->reinject flow. This frees all tags
associated with the packet back to the buffer pool so we can reuse the
completion tags. Since we're saving a copy of the skb pointer and the
BQL info for the reinject completion or timer expire, there's no need to
touch the buffer pool after the miss completion.

For now, allocate the timer info and the xa entry as needed upon
receiving a miss completion. If it fails, we will clean the packet
normally as if it was a normal RS completion. The packet will continue
on the exception path until it is reinjected and put on the wire. The
reinject completion will be ignored. However, because we completed the
packet normally while the packet is actually still going through the
expection path, the BQL will think the bytes went on the wire earlier
than they actually did, which means it will be temporarily inaccruate.

In the reinject->miss flow, we know the descriptors up to the end of the
packet have been read, so we can update next_to_clean. However, we
cannot unmap anything since we don't know the status of the DMA. Set the
status to REINJECT so that when we receive the miss completion (which is
guaranteed so no need for timer), it knows the reinject completion was
already processed.

Signed-off-by: Joshua Hay <joshua.a.hay@intel.com>
Change-type: Other