compat: force drivers to include kcompat.h as the first header
Sometimes, when drivers include kernel headers prior to kcompat.h,
we get a mix of old and new behavior, which is not really supported.
This is strictly necessary for stuff that we override in the COMPAT,
to be prior their kernel counterparts, like the stuff in kcompat_gcc.h
file, but it is also true for the headers that we ship as replacements
of the kernel ones (kcompat_overflow.h, kcompat_cleanup.h).
In the future it may be the case that there will be yet another thing
that needs to be done in a controlled order, and having kcompat.h as
the very first included header helps a lot with such a goal.

This change will also force us to backport stuff into our replacement
files (kcompat_overflow.h, kcompat_cleanup.h) prior to use.

Change-type: FeatureImplementation
Signed-off-by: Przemek Kitszel <przemyslaw.kitszel@intel.com>