compat: remove scoped_cond_guard() and similar
Remove current backports of conditional locks.
We don't use the syntactic sugar for those, and such locks are discouraged
in general, there is no need to make the usage pleasant.
Alternative is to pull even more backports (both in cleanup.h and users),
which is better to be left to when it would be actually needed.

Change-type: ImplementationChange
Signed-off-by: Przemek Kitszel <przemyslaw.kitszel@intel.com>