idpf: avoid bloating &idpf_q_vector with big %NR_CPUS
backport of commit bf9bf7042a38ebd2485592467772db50605bd4a2

With CONFIG_MAXSMP, sizeof(cpumask_t) is 1 Kb. The queue vector
structure has them embedded, which means 1 additional Kb of not
really hotpath data.
We have cpumask_var_t, which is either an embedded cpumask or a pointer
for allocating it dynamically when it's big. Use it instead of plain
cpumasks and put &idpf_q_vector on a good diet.
Also remove redundant pointer to the interrupt name from the structure.
request_irq() saves it and free_irq() returns it on deinit, so that you
can free the memory.

Change-type: EngineeringNotes
Signed-off-by: Alexander Lobakin <aleksander.lobakin@intel.com>
Signed-off-by: Madhu Chittim <madhu.chittim@intel.com>